refactor: put tests in better spot
